requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
distlib==0.3.5
Django==4.0.6
filelock==3.8.0
Flask==2.2.1
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.1.0
MarkupSafe==2.1.1
networkx==2.8.5
nltk==3.7
numpy==1.23.1
pipenv==2022.8.24
platformdirs==2.5.2
pygame==2.1.2
python-decouple==3.6
python-dotenv==0.20.0
regex==2022.7.25
requests==2.28.1
soupsieve==2.3.2.post1
SpeechRecognition==3.8.1
sqlparse==0.4.2
tqdm==4.64.0
urllib3==1.26.10
virtualenv==20.16.3
virtualenv-clone==0.5.7
Werkzeug==2.2.1
zipp==3.8.1
